HXCS-3659 fix imports after renaming files
