+++ lib/core/feature-flags/src/lib/components/flags/flags.component.html
@@ -1,6 +1,6 @@
 <mat-toolbar class="feature-flags-overrides-header">
   <div class="feature-flags-overrides-header-text">
-    <adf-feature-flags-overridability-indicator></adf-feature-flags-overridability-indicator>
+    <adf-feature-flags-override-indicator></adf-feature-flags-override-indicator>
     <span>{{ "CORE.FEATURE-FLAGS.OVERRIDE" | translate }}</span>
   </div>
   <mat-slide-toggle
